I like the idea of nightly commits
So I will keep doing this until I finish this partial re-write. I have
most of the networking and connections done now. Added port class and
object for better organization of port arrays on servers so they are not
just string arrays. Also went back to my original idea of using arouter
for contecting to networks instead of the network class. All that is
left is to fix gameplay classes.