Added a bit storage and router classes
I just had to get this last one in today. I am still working on the
router class a bit. The device needed to be told what its master storage
device was, the one used for logs and system files.

// TODO Tommorow I work on the gamplay package classes, and after that I
will merge these branches. Then I will make a packet class for data
transmission, And after all of that I will make a new branch for the gui
updates.