Ready to merge
All works besides gui but that comes next. Network and service provider
generation works great. I need to comeup with names, brands and models
but that can be done later. The only errors besides gui are in the
gameplay class and I need gui to work  efore I fix them.