Many things added, currently working on an isometric game room
I had a great idea for the game, you can read about it in a block comment in gui/room.java. Now I actually have a reason to use graphics, so users don't just stare at a terminal all the time. I am working on isometric display, windows, terminal changes, rooms, and network and ineternet maps. That will be it for gui changes then and it will let me finnaly work on gameplay.