Starting work on my implementation of GNU CoreUtils
So, figuring out a way to connect a server's filesystem to the real computer's filesystem in order to use an external terminal was going to be way too difficult. I am just going to copy GNU CoreUtils as best as I can. I looked over the list, and it is not too long, since only half of them would even have use in the game.

What says hacking better than linux? When I changed the filesystem to a tree structure, I was basing it off of linux. It makes the storage drive implementation so much easier.

Now I am working on my own implementation of GNU CoreUtils, because they are amazing. Especially if the terminal is going to be a big part of the game, which I plan it will.