Changed the project layout, unfortunately this means a huge diff
I have quite a bit of work done on programs and other gameplay features, but I wanted to fix up the terminal. I am almost done working on the new terminal for the menu, which will be somewhat easily transfered to the ingame one.

It is going to take quite a bit of work to copy the stashed changes to the new project layout.

TL;DR, lots of stashed changes, wanted to get this out of the way.