Updated the menu to use external libraries to handle everything
The lines on the map were untouched, and yet are now somehow broken at this time.

It now uses jline(old version inside of jython, since it messed up with dependencies) as well as jopt simple.

Jline handles the console rendering, and, with some hackery(read as changing functionality without overides), writes it to a output stream to the display label. Jline also handles the code completion. Jopt handles command option parsing.

The next thing up is either copy/paste or copping the code to the in-game terminal. It may be simpler finish the everything, then copy the terminal code. Then again, for copy/paste I would just read the data from the clipboard. Not too sure how to do paste yet, but I have a few ideas.